Also check mimeType
